rename: LGFX_Config to LGFX_Config_Custom
